AllRulesInConfigTest: verify all rules are properly registered
